# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de paiement.",
        "title": "Sikem Payment API",
        "termsOfService": "http://www.sikem.ci/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.sikem.ci/support",
            "email": "support@sikem.ci"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.2"
    },
    "host": "spay.sikem.ci",
    "basePath": "/",
    "paths": {
...
        "/api/users/:id": {
            "get": {
                "description": "Récuperation des informations de l'utilisateur",
                "tags": [
                    "Users"
                ],
                "summary": "Get user data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/users.ResUserAPIGetSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ResFailure"
                        }
                    }
                }
            },
            "put": {
                "description": "Mise à jour des données de l'utilisateur",
                "consumes": [
                    "application/json",
                    "text/xml",
                    "application/x-www-form-urlencoded",
                    "multipart/form-data"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "UpdateInfo user information",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Appliquer la mise à jour au système complet",
                        "name": "apply_for_all_system",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Pays (ex: civ)",
                        "name": "country",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Adresse e-mail valide",
                        "name": "email",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Prénoms",
                        "name": "first_name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Nom",
                        "name": "last_name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Numéro de mobile",
                        "name": "phone_number",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Prefix téléphonique (ex: 225)",
                        "name": "phone_prefix",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/users.ResUserAPIUpdateSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ResFailure"
                        }
                    }
                }
            }
        },
        "/api/users/change-role": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Changement de la permission de l'utilisateur",
                "consumes": [
                    "application/json",
                    "text/xml",
                    "application/x-www-form-urlencoded",
                    "multipart/form-data"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Change user role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Identifiant de connexion de l'utilisateur",
                        "name": "auth_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "maximum": 2,
                        "minimum": 0,
                        "type": "integer",
                        "description": "0: Merchant, 1: Manager, 2: Admin",
                        "name": "role",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/users.ResUserAPIRoleSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ResFailure"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.ErrorAPI": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.ProviderModel": {
            "type": "object",
            "required": [
                "description",
                "health_url",
                "name",
                "pay_check_url",
                "pay_url",
                "support_country"
            ],
            "properties": {
                "asynchrone_mode": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "health_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "name_slug": {
                    "type": "string"
                },
                "pay_check_url": {
                    "type": "string"
                },
                "pay_url": {
                    "type": "string"
                },
                "support_country": {
                    "description": "CIV",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TransactionModel"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.ResFailure": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ErrorAPI"
                    }
                },
                "is_error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "request_date": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "time_elapsed": {
                    "type": "string"
                }
            }
        },
        "models.ServiceModel": {
            "type": "object",
            "required": [
                "description",
                "name"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "country": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "current_amount": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "logo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "name_slug": {
                    "type": "string"
                },
                "permissions": {
                    "description": "Liste de permission des utilisateurs",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ServicePermissionModel"
                    }
                },
                "site_web": {
                    "type": "string"
                },
                "transactions": {
                    "description": "Liste des transaction",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TransactionModel"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.ServicePermissionModel": {
            "type": "object",
            "required": [
                "role"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "role": {
                    "type": "integer"
                },
                "service": {
                    "$ref": "#/definitions/models.ServiceModel"
                },
                "service_id": {
                    "description": "Service",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/models.UserModel"
                },
                "user_id": {
                    "description": "User",
                    "type": "string"
                }
            }
        },
        "models.TransactionModel": {
            "type": "object",
            "required": [
                "amount",
                "amount_with_fee",
                "currency",
                "operation_mode",
                "operation_state",
                "provider",
                "provider_id",
                "reference_id",
                "service",
                "service_id"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "amount_with_fee": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "mode_live": {
                    "type": "boolean"
                },
                "operation_mode": {
                    "description": "CREDIT, DEBIT",
                    "type": "string"
                },
                "operation_msg": {
                    "type": "string"
                },
                "operation_state": {
                    "description": "PENDING, SUCCESS, CANCEL, FAIL",
                    "type": "string"
                },
                "provider": {
                    "$ref": "#/definitions/models.ProviderModel"
                },
                "provider_id": {
                    "description": "Provider",
                    "type": "string"
                },
                "reference_id": {
                    "description": "Réference",
                    "type": "string"
                },
                "service": {
                    "$ref": "#/definitions/models.ServiceModel"
                },
                "service_id": {
                    "description": "Service",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.UserModel": {
            "type": "object",
            "required": [
                "auth_id",
                "first_name",
                "last_name"
            ],
            "properties": {
                "auth_id": {
                    "description": "Identifiant unique keycloak",
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "description": "Information",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "phone_prefix": {
                    "type": "string"
                },
                "role": {
                    "description": "Role",
                    "type": "integer"
                },
                "service_permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ServicePermissionModel"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "users.ResUserAPIGetSuccess": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "user": {
                            "$ref": "#/definitions/models.UserModel"
                        }
                    }
                },
                "is_error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "request_date": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "time_elapsed": {
                    "type": "string"
                }
            }
        },
...
        "users.ResUserAPIRoleSuccess": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "new_role": {
                            "type": "string"
                        },
                        "user": {
                            "$ref": "#/definitions/models.UserModel"
                        }
                    }
                },
                "is_error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "request_date": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "time_elapsed": {
                    "type": "string"
                }
            }
        },
        "users.ResUserAPIUpdateSuccess": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "user": {
                            "$ref": "#/definitions/models.UserModel"
                        }
                    }
                },
                "is_error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "request_date": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "time_elapsed": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}